MIR-647 earthkit-regrid support for linear interpolation from HEALPix orderingConvention=nested
