+++ src/tools/mir-matrix-reorder.cc
@@ -12,7 +12,7 @@
 
 #include <algorithm>
 #include <memory>
-#include <numeric>
+#include <string>
 #include <vector>
 
 #include "eckit/linalg/SparseMatrix.h"
@@ -21,184 +21,26 @@
 #include "eckit/option/FactoryOption.h"
 #include "eckit/option/SimpleOption.h"
 
+#include "mir/reorder/Reorder.h"
 #include "mir/repres/Representation.h"
-#include "mir/repres/proxy/HEALPix.h"
 #include "mir/tools/MIRTool.h"
 #include "mir/util/Exceptions.h"
 #include "mir/util/Log.h"
-#include "mir/util/Mutex.h"
 
 
 namespace mir::tools {
 
 
-using Renumber = std::vector<size_t>;
-
-static util::once_flag ONCE;
-static util::recursive_mutex* MUTEX = nullptr;
-
-class ReorderFactory;
-static std::map<std::string, ReorderFactory*>* M = nullptr;
-
-
-static void init() {
-    MUTEX = new util::recursive_mutex();
-    M     = new std::map<std::string, ReorderFactory*>();
-}
-
-
-struct Reorder {
-    Reorder() = default;
-
-    virtual ~Reorder() = default;
-
-    virtual Renumber reorder() = 0;
-
-    Reorder(const Reorder&)            = delete;
-    Reorder(Reorder&&)                 = delete;
-    Reorder& operator=(const Reorder&) = delete;
-    Reorder& operator=(Reorder&&)      = delete;
-};
-
-
-class ReorderFactory {
-private:
-    std::string name_;
-    virtual Reorder* make(size_t N) = 0;
-
-protected:
-    explicit ReorderFactory(const std::string& name) : name_(name) {
-        util::call_once(ONCE, init);
-        util::lock_guard<util::recursive_mutex> lock(*MUTEX);
-
-        if (M->find(name) == M->end()) {
-            (*M)[name] = this;
-            return;
-        }
-
-        throw exception::SeriousBug("ReorderFactory: duplicated Reorder '" + name + "'");
-    }
-
-    virtual ~ReorderFactory() {
-        util::lock_guard<util::recursive_mutex> lock(*MUTEX);
-        M->erase(name_);
-    }
-
-public:
-    ReorderFactory(const ReorderFactory&)            = delete;
-    ReorderFactory(ReorderFactory&&)                 = delete;
-    ReorderFactory& operator=(const ReorderFactory&) = delete;
-    ReorderFactory& operator=(ReorderFactory&&)      = delete;
-
-    static Reorder* build(const std::string& name, size_t N) {
-        util::call_once(ONCE, init);
-        util::lock_guard<util::recursive_mutex> lock(*MUTEX);
-
-        if (auto j = M->find(name); j != M->end()) {
-            return j->second->make(N);
-        }
-
-        list(Log::error() << "ReorderFactory: unknown '" << name << "', choices are:\n") << std::endl;
-        throw exception::SeriousBug("ReorderFactory: unknown '" + name + "'");
-    }
-
-    static std::ostream& list(std::ostream& out) {
-        util::call_once(ONCE, init);
-        util::lock_guard<util::recursive_mutex> lock(*MUTEX);
-
-        const auto* sep = "";
-        for (const auto& j : *M) {
-            out << sep << j.first;
-            sep = ", ";
-        }
-
-        return out;
-    }
-};
-
-
-template <class T>
-class ReorderBuilder : public ReorderFactory {
-    Reorder* make(size_t N) override { return new T(N); }
-
-public:
-    explicit ReorderBuilder(const std::string& name) : ReorderFactory(name) {}
-};
-
-
-struct Identity final : Reorder {
-    explicit Identity(size_t N) : N_(N) {}
-
-private:
-    Renumber reorder() override {
-        Renumber renumber(N_);
-        std::iota(renumber.begin(), renumber.end(), 0);
-        return renumber;
-    }
-
-    const size_t N_;
-};
-
-
-struct HEALPix : Reorder {
-    explicit HEALPix(size_t N) :
-        N_(N),
-        Nside_([N] {
-            auto Nside = static_cast<size_t>(std::sqrt(static_cast<double>(N) / 12.));
-            ASSERT_MSG(12 * Nside * Nside == N, "Expected N = 12 * Nside ** 2, got N=" + std::to_string(N));
-            return Nside;
-        }()),
-        healpix_(static_cast<int>(Nside_)) {}
-
-    size_t N() const { return N_; }
-    size_t Nside() const { return Nside_; }
-    const repres::proxy::HEALPix::Reorder& healpix() const { return healpix_; }
-
-private:
-    const size_t N_;
-    const size_t Nside_;
-    const repres::proxy::HEALPix::Reorder healpix_;
-};
-
-
-struct HEALPixRingToNested final : HEALPix {
-    explicit HEALPixRingToNested(size_t N) : HEALPix(N) {}
-    Renumber reorder() override {
-        Renumber map(N());
-        for (size_t i = 0; i < N(); ++i) {
-            map[i] = static_cast<size_t>(healpix().ring_to_nest(static_cast<int>(i)));
-        }
-        return map;
-    }
-};
-
-
-struct HEALPixNestedToRing final : HEALPix {
-    explicit HEALPixNestedToRing(size_t N) : HEALPix(N) {}
-    Renumber reorder() override {
-        Renumber map(N());
-        for (size_t i = 0; i < N(); ++i) {
-            map[i] = static_cast<size_t>(healpix().nest_to_ring(static_cast<int>(i)));
-        }
-        return map;
-    }
-};
-
-
-static const std::string IDENTITY = "identity";
-
-static const ReorderBuilder<Identity> __reorder1(IDENTITY);
-static const ReorderBuilder<HEALPixRingToNested> __reorder2("healpix-ring-to-nested");
-static const ReorderBuilder<HEALPixNestedToRing> __reorder3("healpix-nested-to-ring");
+using Reorder = reorder::Reorder;
 
 
 struct MIRMatrixReorder : MIRTool {
     MIRMatrixReorder(int argc, char** argv) : MIRTool(argc, argv) {
-        using eckit::option::FactoryOption;
-        using eckit::option::SimpleOption;
-        options_.push_back(new FactoryOption<ReorderFactory>("reorder-rows", "Reordering rows method", IDENTITY));
-        options_.push_back(new FactoryOption<ReorderFactory>("reorder-cols", "Reordering columns method", IDENTITY));
-        options_.push_back(new SimpleOption<bool>("transpose", "Transpose matrix", false));
+        options_.push_back(new eckit::option::FactoryOption<reorder::ReorderFactory>(
+            "reorder-rows", "Reordering rows method", "identity"));
+        options_.push_back(new eckit::option::FactoryOption<reorder::ReorderFactory>(
+            "reorder-cols", "Reordering columns method", "identity"));
+        options_.push_back(new eckit::option::SimpleOption<bool>("transpose", "Transpose matrix", false));
     }
 
     int numberOfPositionalArguments() const override { return 2; }
@@ -225,10 +67,12 @@ void MIRMatrixReorder::execute(const eckit::option::CmdArgs& args) {
 
 
     // renumbering maps
-    auto rows = std::unique_ptr<Reorder>(ReorderFactory::build(args.getString("reorder-rows"), M.rows()))->reorder();
+    auto rows =
+        std::unique_ptr<Reorder>(reorder::ReorderFactory::build(args.getString("reorder-rows"), M.rows()))->reorder();
     ASSERT(rows.size() == M.rows());
 
-    auto cols = std::unique_ptr<Reorder>(ReorderFactory::build(args.getString("reorder-cols"), M.cols()))->reorder();
+    auto cols =
+        std::unique_ptr<Reorder>(reorder::ReorderFactory::build(args.getString("reorder-cols"), M.cols()))->reorder();
     ASSERT(cols.size() == M.cols());
